Setup foreman
This commit sets up foreman.
The foreman file is added to tell foreman which port to operate on.
The Procfile is added to tell foreman how to start up.
The values in `.env` can be overriden in `.env.local` which is not tracked.

After this commit, the app should be launched with `foreman start -e .evn,.env.local`
or with the start script: `./bin/start`.