Setup Devise and add User model
This commit sets up devise.

It adds the gem and does some configurations.

The dotenv gem was added so the secret key is available without having to run foreman.

A User model was added.
This model has a pulymorphic "role" relationship which will be used to relate it to the user profile
models that exist already (Teacher, Director, Student).