Replace refile with bare AWS
We had some performance issues with Refile given the size of the images we were uploading. This commit replaces it with a new custom photo processing backend.

The new photo processing backend serves pre-signed upload URLs that enable photos to be uploaded directly to S3 from the frontend. At that point, instead of being processed by the web application, they are added to an SQS queue for a separate photo processing service to process. While the photos are being processed, the original unprocessed image is rendered for requests related to that image.

This should significantly reduce the load on the web application during photo upload realizing major improvements in performance.